backend_todo/api/views.py
from rest_framework import generics, serializers
from .models import Todo
from .serializer import TodoSerializer
# Create your views here.

class TodoGetCreate(generics.ListCreateAPIView):
    queryset = Todo.objects.all()
    serializer_class = TodoSerializer

class TodoUpdateDelete(generics.RetrieveUpdateDestroyAPIView):
    queryset = Todo.objects.all()
    serializer_class = TodoSerializer



